open files in actual JSON editor
