+++ renderer.js
@@ -1,5 +1,6 @@
 const {app, BrowserWindow, Menu} = require('electron')
-const { FeatureLayerEditor, BaseLayerEditor } = require('./lib/mapviewer/LayerEditor')
+const { BaseLayerEditor } = require('./lib/mapviewer/BaseLayerEditor')
+const { FeatureLayerEditor } = require('./lib/mapviewer/FeatureLayerEditor')
 const path = require('path')
 const url = require('url')
 const fs = require('fs')
@@ -158,7 +159,8 @@ fs.readFile('config/menutemplate.json', (err, data) => {
 let createWindow = () => {
   win = new BrowserWindow({
     'width': 1000,
-    'height': 800
+    'height': 800,
+    'show': false
   })
   win.loadURL(url.format({
     pathname: path.join(__dirname, 'index.html'),
@@ -166,12 +168,17 @@ let createWindow = () => {
     slashes: true
   }))
 
+  win.once('ready-to-show', () => {
+    win.show()
+  })
+
   win.on('closed', () => {
     win = null
   })
 }
 
 app.on('ready', createWindow)
+
 app.on('window-all-closed', () => {
   if (process.platform !== 'darwin') {
     app.quit()
